Move REST controllers to separate package
