+++ wm-system/src/main/kotlin/com/ajaxproject/warehouse/controller/rest/CustomersController.kt
@@ -1,4 +1,4 @@
-package com.ajaxproject.warehouse.controller
+package com.ajaxproject.warehouse.controller.rest
 
 import com.ajaxproject.warehouse.dto.CustomerCreateDto
 import com.ajaxproject.warehouse.dto.CustomerDataDto
